# src/swagger/swagger_output.yaml
...
openapi: 3.0.0
info:
  version: 0.0.0
  title: Vinymatic API
  description: ""
host: localhost:5000
basePath: ""
...
schemes: []
securityDefinitions: {}
consumes: []
produces: []
paths:
...
  "/artists":
    get:
      tags:
      - Artist
      description: ""
      parameters:
      - $ref: "#/components/parameters/pageParam"
      - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artists"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    post:
      tags:
      - Artist
      description: ""
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddArtist"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/artists/search":
    get:
      tags:
      - Artist
      - Search
      description: ""
      parameters: 
      - $ref: "#/components/parameters/searchField"
      - $ref: "#/components/parameters/searchValue"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artists"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/artist/{id}":
    get:
      tags:
      - Artist
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    patch:
      tags:
      - Artist
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddArtist"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "304":
          description: Not Modified
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/artist/{id}/discography":
    get:
      tags:
      - Artist
      description: ""
      parameters:
      - name: counting
        in: query
        schema:
          type: boolean
      - name: groupBy
        in: query
        description: "Valeurs possibles: __albums__, __singles__ et __compilations__" 
        schema:
          type: string
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
# Audio ✔︎
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
    searchField:  
      name: withCol[field]
      description: Permet de rechercher sur une colomne en particulier avec le __nom de colomne__( ___field___ ) souhaité
      example: name
      in: query
      required: true
      schema:
        type: array
        items:
          type: string
    searchValue:  
      name: withCol[value]
      description: Permet de rechercher sur une colomne en particulier avec la __valeur__( ___value___ ) souhaitée
      example: michael
      in: query
      required: true
      schema:
        type: array
        items:
          type: string
...
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  requestBodies:
...
    AddArtist:
      type: object
      properties:
        name:
          type: string
        fullName:
          type: string
        aliasNames:
          type: array
        variantNames:
          type: array
        description:
          type: text
        inGroups:
          type: array
        images:
          type: array
          items: 
            type: string
            format: binary
        thumbnail:
          type: string
          format: binary
      # required:
      # - name
...
  schemas:
...
    MinOfArtist:
      type: object
      properties:
        id: 
          type: integer
        idArtist: 
          type: integer
        name:
          type: string
        thumbnail:
          type: string
        artistUrl:
          type: string
        ressourceUrl:
          type: string
    MinOfAudio:
      type: object
      properties:
        id: 
          type: integer
        title: 
          type: string
        image: 
          type: string
        duration: 
          type: integer
        audioUrl: 
          type: string
        resourceUrl: 
          type: string
...
    MinOfMaster:
      type: object
      properties:
        id: 
          type: integer
        idMaster:
          type: integer  
        title:
          type: string
        releaseDate:
          type: string    
        thumbnail:
          type: string
        masterUrl:
          type: string
        resourceUrl:
          type: string
...
    MinOfVinyl:
      type: object
      properties:
        id: 
          type: integer
        idRelease: 
          type: integer  
        title:
          type: string  
        thumbnail:
          type: string  
        vinylUrl:
          type: string
        resourceUrl:
          type: string
  
  # Array of
    ArrayOfArtist:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfArtist"
    ArrayOfAudio:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfAudio"
...
    ArrayOfMaster:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfMaster"
...
    ArrayOfVinyl:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfVinyl"
...
    Artist:
      type: object
      properties: 
        artist:
          allOf:
          - $ref: "#/components/schemas/MinOfArtist"
          - type: object
            properties:
              fullName:
                type: string
              aliasNames:
                type: array
                items:
                  type: string
              variantNames:
                type: array
                items:
                  type: string
              description:
                type: text
              inGroups:
                type: array
                items:
                  type: string
              images:
                type: array
                items:
                  type: string
              VinylMainArtists: 
                allOf:
                - $ref: "#/components/schemas/ArrayOfVinyl"
              VinylCredits:
                allOf:
                - $ref: "#/components/schemas/ArrayOfVinyl"
              MasterMainArtists: 
                allOf:
                - $ref: "#/components/schemas/ArrayOfMaster"
              MasterCredits:
                allOf:
                - $ref: "#/components/schemas/ArrayOfMaster"
              Audio: 
                allOf:
                - $ref: "#/components/schemas/ArrayOfAudio"
...
    Artists:
      type: object
      properties:
        artists:
          allOf:
          - $ref: "#/components/schemas/ArrayOfArtist"
...